pyproject.toml
[tool.poetry]
name = "jenkinsctl"
version = "0.2.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
api4jenkins = "^2.0.3"
pyyaml = "^6.0.2"
dynaconf = "^3.2.6"
argcomplete = "^3.5.0"
rich = "^13.8.1"
python-dateutil = "^2.9.0.post0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
jenkinsctl = "jenkinsctl.cli:get_args"
